Fix DQN-CartPole. Add trained weights.

+++ Deep-Q-Networks/CartPole/ReplayMemory.py
@@ -18,10 +18,10 @@ class ReplayMemory:
         else:
             samp = random.sample(self.replay_buffer, batch_size)
             obss, actions, rewards, next_obss, terminateds = zip(*samp)
-            return (torch.tensor(np.array(obss)),
+            return (torch.tensor(np.array(obss)).float(),
                     torch.tensor(np.array(actions)),
-                    torch.tensor(np.array(rewards)),
-                    torch.tensor(np.array(next_obss)),
+                    torch.tensor(np.array(rewards)).float(),
+                    torch.tensor(np.array(next_obss)).float(),
                     torch.tensor(np.array(terminateds)))
     
     def __len__(self):
